Fallback logic for streams not on Firestore
